# backend/api-gateway-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Commit API Gateway
  description: |
    HTTP surface exposed by the **api-gateway** service (`services/api-gateway`).

    - **Auth** paths under `/api/v1/auth/*` are reverse-proxied to the authentication service (FastAPI).
      Request/response bodies match that service unless noted.
    - **Profiles**, **recommendations**, **swipes**, and **matches** are routed through the gateway to the backing gRPC services. **GET /swipes** and **GET /recommendations** both resolve user ids from their backing service, then load rows via **profiles** `GetProfiles`.
    - **Chats** (`/api/v1/chats/*`) are reverse-proxied to the **chats** FastAPI service. The gateway validates JWT and forwards **`X-User-Id`** (numeric user id); the chats service does not verify JWT itself.
    - **Recommendations** HTTP routes call the **recommendations** gRPC service for candidate ids (and filter CRUD on `/recommendations/filters`); **GET /recommendations** loads profile rows via **profiles** `GetProfiles` by those ids (except **compatibility**, which stays inside the recommendations service and uses **Ollama** for text generation).
    - Protected routes require `Authorization: Bearer <JWT>`. The gateway validates the token and expects a numeric `sub` claim (user id).

    **JSON naming**
    - Profile handlers serialize protobuf types with Go `encoding/json` using **snake_case** field names (e.g. `profile_data`, `user_id`).
...

...
  version: 1.0.0
servers:
...
paths:
...
  /api/v1/chats:
    get:
      tags: [Chats]
      summary: List chats for the current user
      description: |
        Proxied to `GET /api/v1/chats` on the chats service. Returns chat summaries sorted by latest activity.
      operationId: chatsList
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK — JSON array of chat summaries (**snake_case** from FastAPI)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSummaryItem'
        '401':
          description: Unauthorized (missing JWT at gateway or missing `X-User-Id` at chats service)
        '502':
          description: Bad gateway (chats service unreachable)
  /api/v1/chats/{peer_user_id}:
    parameters:
      - name: peer_user_id
        in: path
        required: true
        description: Other participant user id (the chat partner)
        schema:
          type: integer
          format: int64
    get:
      tags: [Chats]
      summary: Open chat — list messages with a user
      description: |
        Proxied to `GET /api/v1/chats/{peer_user_id}`. Returns all messages if a chat exists between the JWT user and `peer_user_id`.
      operationId: chatsGetWithPeer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatWithMessagesResponse'
        '401':
          description: Unauthorized
        '404':
          description: Chat not found for this pair
        '502':
          description: Bad gateway
    delete:
      tags: [Chats]
      summary: Delete chat with a user
      description: |
        Proxied to `DELETE /api/v1/chats/{peer_user_id}`. Removes the chat (and related data) for both participants.
      operationId: chatsDeleteWithPeer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted (no body)
        '401':
          description: Unauthorized
        '502':
          description: Bad gateway
  /api/v1/chats/{peer_user_id}/messages:
    parameters:
      - name: peer_user_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Chats]
      summary: Send a chat message (text and/or image)
      description: |
        Proxied to `POST /api/v1/chats/{peer_user_id}/messages` on the chats service (`201 Created`).

        Use **`application/json`** with `{"body":"..."}` for a plain text message, or **`multipart/form-data`**
        with optional part **`body`** (caption) and/or **`file`** (image). At least one of `body` or `file` is required
        for multipart. Images must be **JPEG, PNG, or WebP** (max **10 MiB**); the service stores the object in MinIO
        and returns `image_storage_key` on the created message. Kafka `chat.message` preview uses the caption when
        present, otherwise `"[image]"` for image-only messages.
      operationId: chatsSendMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendChatTextMessageBody'
          multipart/form-data:
            schema:
              type: object
              properties:
                body:
                  type: string
                  maxLength: 10000
                  description: Optional caption when sending an image, or text-only content when `file` is omitted.
                file:
                  type: string
                  format: binary
                  description: Optional image (jpeg, png, webp); max 10 MiB.
              description: |
                Send at least one of `body` or `file`. Both may be set (image with caption).
      responses:
        '201':
          description: Created — message object (**snake_case**)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageItem'
        '400':
          description: |
            Invalid request (e.g. neither `body` nor `file`, empty JSON `body`, unsupported image type, image too large,
            or invalid peer).
        '401':
          description: Unauthorized
        '404':
          description: Chat not found for this pair
        '422':
          description: Invalid JSON body
        '502':
          description: Bad gateway (upstream chats service or object storage failure)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RSA-signed JWT; `sub` claim must be numeric user id.
  schemas:
...
    ChatSummaryItem:
      type: object
      description: One row from `GET /api/v1/chats` (chats service)
      properties:
        chat_id:
          type: string
          description: Chat UUID as string
        peer_user_id:
          type: integer
          format: int64
        last_message_at:
          type: string
          nullable: true
          description: ISO-8601 timestamp or null if no messages
        last_message_preview:
          type: string
          nullable: true
          description: Text preview or `[image]` for image-only last message
    ChatMessageItem:
      type: object
      properties:
        id:
          type: string
          description: Message UUID as string
        sender_id:
          type: integer
          format: int64
        body:
          type: string
          nullable: true
        image_storage_key:
          type: string
          nullable: true
        created_at:
          type: string
          description: ISO-8601 timestamp
    ChatWithMessagesResponse:
      type: object
      properties:
        chat_id:
          type: string
        peer_user_id:
          type: integer
          format: int64
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageItem'
    SendChatTextMessageBody:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
          description: Message text